Everything has been reframed in terms of masking. Now there are far less compatibility issues depending on the Tensorflow version. New examples are given.
