+++ spectraltools/__init__.py
@@ -1,2 +1,2 @@
 from .spectraldense import Spectral
-from .spectralprune import spectral_pruning, spectral_pretrain, spectral_distiller
+from .spectralprune import prune_percentile, metric_based_pruning, original_model
